% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_freq}
\alias{check_freq}
\title{Check frequency code}
\usage{
check_freq(trade_direction)
}
\arguments{
\item{trade_direction}{A character string specifying the frequency of the data. Must be one of "A", "Q", or "M".}
}
\value{
A character string specifying the frequency of the data.
}
\description{
Check frequency code
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_freq("A") # returns "A"
check_freq("Q") # returns "Q"
check_freq("M") # returns "M"
check_freq("D") # throws an error because "D" is not a valid frequency code
\dontshow{\}) # examplesIf}
}
