% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_period}
\alias{check_period}
\title{Check if period is valid and return a comma-separated string of periods}
\usage{
check_period(period)
}
\arguments{
\item{period}{A period reference as an integer or character vector or a range as "start:end".}
}
\value{
A comma-separated string of periods.
}
\description{
Check if period is valid and return a comma-separated string of periods
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_period(1999:2002)
check_period(c('1999', '2020'))
check_period('2021')
\dontshow{\}) # examplesIf}
}
