% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS}
\alias{HS}
\title{HS Codes for that data}
\format{
## `HS`
A data frame with 8,262 rows and 3 columns:
\describe{
  \item{id}{ID used for querying}
  \item{text}{Description of Code}
  \item{parent}{parent id node}
  ...
}
}
\source{
<https://comtradeapi.un.org/files/v1/app/reference/HS.json>
}
\usage{
HS
}
\description{
From the UN Comtrade API
}
\keyword{datasets}
