% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade_requests.R
\name{perform_comtrade_request}
\alias{perform_comtrade_request}
\title{make_comtrade_request}
\usage{
perform_comtrade_request(req)
}
\arguments{
\item{req}{a valid comtrade request built by the build_comtrade_request function}
}
\value{
json data from comtrade and possible error codes
}
\description{
make_comtrade_request
}
