% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade_requests.R
\name{process_comtrade_response}
\alias{process_comtrade_response}
\title{process_comtrade_response
The function adds the respective iso codes for the reporter and partner countries, as well as the commodity code description.}
\usage{
process_comtrade_response(resp)
}
\arguments{
\item{resp}{a valid httr2 response object created from the function `perform_comtrade_request`}
}
\value{
a data.frame object with the results
}
\description{
process_comtrade_response
The function adds the respective iso codes for the reporter and partner countries, as well as the commodity code description.
}
