% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REPORTER}
\alias{REPORTER}
\title{REPORTER Codes for that data}
\format{
## `REPORTER`
A data frame with 8,262 rows and 3 columns:
\describe{
  \item{id}{ID used for querying}
  \item{text}{Description of Code}
  \item{isGroup}{whether it is not a country, but a group reference}
  ...
}
}
\source{
<https://comtradeapi.un.org/files/v1/app/reference/Reporters.json>
}
\usage{
REPORTER
}
\description{
From the UN Comtrade API
}
\keyword{datasets}
