% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_primary_comtrade_key}
\alias{set_primary_comtrade_key}
\title{set_primary_comtrade_key}
\usage{
set_primary_comtrade_key(key = NULL)
}
\arguments{
\item{key}{Provide your primary comtrade key}
}
\value{
Saves your comtrade primary key in the environment.
}
\description{
set_primary_comtrade_key
}
